+++ src/core/deals/MidaBrokerDealUtilities.ts
@@ -1,0 +1,60 @@
+/*
+ * Copyright Reiryoku Technologies and its contributors, https://www.reiryoku.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+*/
+
+import { MidaBrokerDeal } from "#deals/MidaBrokerDeal";
+import { MidaBrokerOrder } from "#orders/MidaBrokerOrder";
+
+export namespace MidaBrokerDealUtilities {
+    export function filterExecutedDeals (deals: MidaBrokerDeal[]): MidaBrokerDeal[] {
+        const executedDeals: MidaBrokerDeal[] = [];
+
+        for (const deal of deals) {
+            if (deal.isExecuted) {
+                executedDeals.push(deal);
+            }
+        }
+
+        return executedDeals;
+    }
+
+    export function filterRejectedDeals (deals: MidaBrokerDeal[]): MidaBrokerDeal[] {
+        const rejectedDeals: MidaBrokerDeal[] = [];
+
+        for (const deal of deals) {
+            if (deal.isRejected) {
+                rejectedDeals.push(deal);
+            }
+        }
+
+        return rejectedDeals;
+    }
+
+    export function getDealsFromOrders (orders: MidaBrokerOrder[]): MidaBrokerDeal[] {
+        const deals: MidaBrokerDeal[] = [];
+
+        for (const order of orders) {
+            deals.push(...order.deals);
+        }
+
+        return deals;
+    }
+}
