+++ src/core/orders/MidaBrokerOrderUtilities.ts
@@ -1,0 +1,50 @@
+/*
+ * Copyright Reiryoku Technologies and its contributors, https://www.reiryoku.com
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+*/
+
+import { MidaBrokerOrder } from "#orders/MidaBrokerOrder";
+import { MidaBrokerOrderStatus } from "#orders/MidaBrokerOrderStatus";
+
+export namespace MidaBrokerOrderUtilities {
+    export function filterPendingOrders (orders: MidaBrokerOrder[]): MidaBrokerOrder[] {
+        const pendingOrders: MidaBrokerOrder[] = [];
+
+        for (const order of orders) {
+            if (order.status === MidaBrokerOrderStatus.PENDING) {
+                pendingOrders.push(order);
+            }
+        }
+
+        return pendingOrders;
+    }
+
+    export function filterExecutedOrders (orders: MidaBrokerOrder[]): MidaBrokerOrder[] {
+        const executedOrders: MidaBrokerOrder[] = [];
+
+        for (const order of orders) {
+            if (order.isExecuted) {
+                executedOrders.push(order);
+            }
+        }
+
+        return executedOrders;
+    }
+}
